@@ -1,7 +1,8 @@
-from typing import Iterator
+from collections.abc import Iterator
 
+import numpy as np
 import sklearn
-from sklearn.model_selection import KFold, GroupKFold, GroupShuffleSplit
+from sklearn.model_selection import GroupKFold, GroupShuffleSplit, KFold
 
 from lir.data.models import DataSet, DataStrategy
 
@@ -16,31 +17,17 @@ class BinaryTrainTestSplit(DataStrategy):
         self.source = source
         self.test_size = test_size
         self.seed = seed
-        self.shuffle = True if self.seed is not None else False
 
     def __iter__(self) -> Iterator:
         """Allow iteration by looping over the resulting train/test split(s)."""
-        (
-            features_train,
-            features_test,
-            labels_train,
-            labels_test,
-            meta_train,
-            meta_test,
-        ) = sklearn.model_selection.train_test_split(
-            *self.source.get_instances(),
-            test_size=self.test_size,
-            shuffle=self.shuffle,
-            random_state=self.seed,
+        instances = self.source.get_instances()
+
+        indexes = np.arange(len(instances))
+        indexes_train, indexes_test = sklearn.model_selection.train_test_split(
+            indexes, stratify=instances.labels, test_size=self.test_size, shuffle=True, random_state=self.seed
         )
-        yield (
-            (features_train, labels_train, meta_train),
-            (
-                features_test,
-                labels_test,
-                meta_test,
-            ),
-        )
+
+        yield instances[indexes_train], instances[indexes_test]
 
 
 class BinaryCrossValidation(DataStrategy):
@@ -53,21 +40,14 @@ class BinaryCrossValidation(DataStrategy):
         self.source = source
         self.folds = folds
         self.seed = seed
-        self.shuffle = True if self.seed is not None else False
+        self.shuffle = True if self.seed is not None else False  # noqa: SIM210
 
     def __iter__(self) -> Iterator:
         """Allow iteration by looping over the resulting train/test split(s)."""
         kf = KFold(n_splits=self.folds, shuffle=self.shuffle, random_state=self.seed)
-        features, labels, meta = self.source.get_instances()
-        for i, (train_index, test_index) in enumerate(kf.split(features, y=labels)):
-            yield (
-                (features[train_index], labels[train_index], meta[train_index]),
-                (
-                    features[test_index],
-                    labels[test_index],
-                    meta[test_index],
-                ),
-            )
+        instances = self.source.get_instances()
+        for _i, (train_index, test_index) in enumerate(kf.split(instances.features, y=instances.labels)):
+            yield instances[train_index], instances[test_index]
 
 
 class MulticlassTrainTestSplit(DataStrategy):
@@ -84,12 +64,10 @@ class MulticlassTrainTestSplit(DataStrategy):
     def __iter__(self) -> Iterator:
         """Allow iteration by looping over the resulting train/test split(s)."""
         splitter = GroupShuffleSplit(test_size=self.test_size, n_splits=1, random_state=self.seed)
-        features, labels, meta = self.source.get_instances()
-        ((train_index, test_index),) = splitter.split(features, labels, labels)
+        instances = self.source.get_instances()
+        ((train_index, test_index),) = splitter.split(instances.features, instances.labels, instances.labels)
 
-        training_set = features[train_index], labels[train_index], meta[train_index]
-        test_set = features[test_index], labels[test_index], meta[test_index]
-        yield training_set, test_set
+        yield instances[train_index], instances[test_index]
 
 
 class MulticlassCrossValidation(DataStrategy):
@@ -106,13 +84,6 @@ class MulticlassCrossValidation(DataStrategy):
         """Allow iteration by looping over the resulting train/test split(s)."""
         kf = GroupKFold(n_splits=self.folds)
 
-        features, labels, meta = self.source.get_instances()
-        for i, (train_index, test_index) in enumerate(kf.split(features, groups=labels)):
-            yield (
-                (features[train_index], labels[train_index], meta[train_index]),
-                (
-                    features[test_index],
-                    labels[test_index],
-                    meta[test_index],
-                ),
-            )
+        instances = self.source.get_instances()
+        for _i, (train_index, test_index) in enumerate(kf.split(instances.features, groups=instances.labels)):
+            yield instances[train_index], instances[test_index]
