@@ -1,4 +1,5 @@
 from collections.abc import Callable
+
 import numpy as np
 import sklearn
 from scipy.interpolate import interp1d
@@ -38,21 +39,21 @@ class PercentileRankTransformer(sklearn.base.TransformerMixin):
     def __init__(self) -> None:
         self.rank_functions: list[Callable] | None = None
 
-    def fit(self, X: np.ndarray, y: np.ndarray | None = None) -> "PercentileRankTransformer":
+    def fit(self, X: np.ndarray, y: np.ndarray | None = None) -> 'PercentileRankTransformer':
         X = X.reshape(X.shape[0], -1)
-        ranks_X = rankdata(X, method="max", axis=0) / X.shape[0]
+        ranks_X = rankdata(X, method='max', axis=0) / X.shape[0]
         self.rank_functions = [
             interp1d(X[:, i], ranks_X[:, i], bounds_error=False, fill_value=(0, 1)) for i in range(X.shape[1])
         ]
         return self
 
     def transform(self, X: np.ndarray) -> np.ndarray:
-        assert self.rank_functions, "transform() called before fit()"
+        assert self.rank_functions, 'transform() called before fit()'
         original_shape = X.shape
         X = X.reshape(X.shape[0], -1)
         assert X.shape[1] == len(self.rank_functions), (
-            f"number of features {X.shape[1]} does not match "
-            "the number of features {len(self.rank_functions)} used for fit()"
+            f'number of features {X.shape[1]} does not match '
+            'the number of features {len(self.rank_functions)} used for fit()'
         )
         ranks = [self.rank_functions[i](X[:, i]) for i in range(X.shape[1])]
         return np.stack(ranks, axis=1).reshape(*original_shape)
