@@ -1,8 +1,7 @@
 from pathlib import Path
 from typing import Any
 
-
-from lir.config.base import ConfigParser, pop_field, ContextAwareDict
+from lir.config.base import ConfigParser, ContextAwareDict, pop_field
 from lir.config.transform import parse_module
 from lir.transform import Tee
 
@@ -14,7 +13,7 @@ class TeeParser(ConfigParser):
         output_dir: Path,
     ) -> Any:
         transformers = []
-        modules = pop_field(config, "modules")
+        modules = pop_field(config, 'modules')
         for module_config in modules:
             transformers.append(parse_module(module_config, output_dir, module_config.context))
 
