@@ -6,6 +6,7 @@ import numpy as np
 from lir.config.base import config_parser, pop_field
 from lir.config.substitution import ContextAwareDict
 from lir.data.models import DataSet
+from lir.lrsystems.lrsystems import FeatureData
 
 
 class SynthesizedDimension(NamedTuple):
@@ -43,7 +44,7 @@ class SynthesizedNormalMulticlassData(DataSet):
         measurements = np.concatenate([population] * self.sources_size) + measurement_error
         return measurements
 
-    def get_instances(self) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
+    def get_instances(self) -> FeatureData:
         """
         Return instances with randomly synthesized data and multi-class labels.
 
@@ -55,30 +56,29 @@ class SynthesizedNormalMulticlassData(DataSet):
         measurements = [self._generate_dimension(rng, dim) for dim in self.dimensions]
         measurements = np.stack(measurements, axis=1)
         labels = np.concatenate([np.arange(self.population_size)] * self.sources_size)
-        meta = np.zeros((measurements.shape[0], 0))
 
-        return measurements, labels, meta
+        return FeatureData(features=measurements, labels=labels)
 
 
 @config_parser
 def synthesized_normal_multiclass(config: ContextAwareDict, _: Path) -> DataSet:
     """Set up (multiple class) data source class to obtain normally distributed data from configuration."""
-    seed = pop_field(config, "seed", validate=int, required=False)
+    seed = pop_field(config, 'seed', validate=int, required=False)
 
-    population = pop_field(config, "population")
-    population_size = pop_field(population, "size", validate=int)
+    population = pop_field(config, 'population')
+    population_size = pop_field(population, 'size', validate=int)
     instances_per_source = pop_field(
         population,
-        "instances_per_source",
+        'instances_per_source',
         validate=int,
     )
 
-    dimensions_cfg = pop_field(config, "dimensions")
+    dimensions_cfg = pop_field(config, 'dimensions')
     dimensions = []
     for dim in dimensions_cfg:
-        mean = pop_field(dim, "mean", validate=float)
-        std = pop_field(dim, "std", validate=float)
-        error_std = pop_field(dim, "error_std", validate=float)
+        mean = pop_field(dim, 'mean', validate=float)
+        std = pop_field(dim, 'std', validate=float)
+        error_std = pop_field(dim, 'error_std', validate=float)
         dimensions.append(SynthesizedDimension(mean, std, error_std))
 
     return SynthesizedNormalMulticlassData(
