@@ -1,15 +1,14 @@
 import logging
 from abc import ABC, abstractmethod
-from collections.abc import Iterable, Sequence, Callable
+from collections.abc import Iterable, Sequence
 from pathlib import Path
 
-import numpy as np
 from tqdm import tqdm
 
 import lir
-from lir.aggregation import Aggregation
-from lir.data.models import DataStrategy
-from lir.lrsystems.lrsystems import LRSystem
+from lir.aggregation import Aggregation, AggregationData
+from lir.data.models import DataStrategy, concatenate_instances
+from lir.lrsystems.lrsystems import LLRData, LRSystem
 
 
 LOG = logging.getLogger(__name__)
@@ -22,17 +21,15 @@ class Experiment(ABC):
         self,
         name: str,
         data: DataStrategy,
-        aggregations: Sequence[Aggregation],
-        visualization_functions: list[Callable],
+        outputs: Sequence[Aggregation],
         output_path: Path,
     ):
         self.name = name
         self.data = data
-        self.aggregations = aggregations
-        self.visualization_functions = visualization_functions
+        self.outputs = outputs
         self.output_path = output_path
 
-    def _run_lrsystem(self, lrsystem: LRSystem) -> tuple[np.ndarray, np.ndarray | None]:
+    def _run_lrsystem(self, lrsystem: LRSystem) -> LLRData:
         """Run experiment on a single LR system configuration using the provided data(setup).
 
         First, the data is split into a training and testing subset, according to the provided
@@ -45,39 +42,28 @@ class Experiment(ABC):
         through the `visualization_functions` and stored in the `output_path` directory.
         """
         # Placeholders for numpy array's of LLRs and labels obtained from each train/test split
-        llrs = []
-        labels = []
+        llr_sets: list[LLRData] = []
 
         # Split the data into a train / test subset, according to the provided DataSetup. This could
         # for example be a simple binary split or a multiple fold cross validation split.
-        for (features_train, labels_train, meta_train), (
-            features_test,
-            labels_test,
-            meta_test,
-        ) in self.data:
-            lrsystem.fit(features_train, labels_train, meta_train)
-            subset_llrs, subset_labels, subset_meta = lrsystem.apply(features_test, labels_test, meta_test)
+        for training_data, test_data in self.data:
+            lrsystem.fit(training_data)
+            subset_llr_results: LLRData = lrsystem.apply(test_data)
+
             # Store results (numpy arrays) into the placeholder lists
-            llrs.append(subset_llrs)
-            if subset_labels is not None:
-                labels.append(subset_labels)
+            llr_sets.append(subset_llr_results)
 
         # Combine collected numpy array's after iteration over the train/test split(s)
-        llrs = np.concatenate(llrs)
-        labels = np.concatenate(labels) if labels else None
+        combined_llrs: LLRData = concatenate_instances(*llr_sets)
 
-        # Generate visualization output as configured by `visualization_functions`
-        # and write graphical output to the `output_path`.
+        # Generate output as configured by `outputs` and write these output to
+        # the `output_path`.
         output_dir = self.output_path / lrsystem.name
-        LOG.debug(f"writing visualizations to {output_dir}")
-        for visualization_function in self.visualization_functions:
-            visualization_function(output_dir, llrs, labels)
+        LOG.debug(f'writing outputs to {output_dir}')
+        for output in self.outputs:
+            output.report(AggregationData(llrdata=combined_llrs, lrsystem=lrsystem, parameters=lrsystem.parameters))
 
-        # Construct a `results` dictionary of metrics indicating the performance of the given LR system
-        for aggregation in self.aggregations:
-            aggregation.report(llrs, labels, lrsystem.parameters)
-
-        return llrs, labels
+        return combined_llrs
 
     @abstractmethod
     def _generate_and_run(self) -> None:
@@ -93,8 +79,8 @@ class Experiment(ABC):
         try:
             self._generate_and_run()
         finally:
-            for aggregation in self.aggregations:
-                aggregation.close()
+            for output in self.outputs:
+                output.close()
 
 
 class PredefinedExperiment(Experiment):
@@ -104,12 +90,11 @@ class PredefinedExperiment(Experiment):
         self,
         name: str,
         data: DataStrategy,
-        aggregations: Sequence[Aggregation],
-        visualization_functions: list[Callable],
+        outputs: Sequence[Aggregation],
         output_path: Path,
         lrsystems: Iterable[LRSystem],
     ):
-        super().__init__(name, data, aggregations, visualization_functions, output_path)
+        super().__init__(name, data, outputs, output_path)
         self.lrsystems = lrsystems
 
     def _generate_and_run(self) -> None:
