@@ -1,33 +1,33 @@
-import numpy as np
+from lir import Transformer
+from lir.lrsystems.lrsystems import FeatureData, LLRData, LRSystem
 
-from lir.lrsystems.lrsystems import LRSystem, Pipeline
 
+class BinaryLRSystem(LRSystem):
+    """
+    LR system for binary data and a linear pipeline.
 
-class SpecificSourceSystem(LRSystem):
-    """Representation of a specific source, feature based LR system.
+    This may be used in specific source feature based LR systems.
 
     In this strategy, a set of instances - captured within the
     feature vector X - and a set of (ground-truth) labels are used to train and
     afterward calculate corresponding LLR's for given feature vectors.
     """
 
-    def __init__(self, name: str, pipeline: Pipeline):
+    def __init__(self, name: str, pipeline: Transformer):
         super().__init__(name)
         self.pipeline = pipeline
 
-    def fit(self, instances: np.ndarray, labels: np.ndarray, meta: np.ndarray) -> "LRSystem":
-        self.pipeline.fit(instances, labels)
+    def fit(self, instances: FeatureData) -> 'LRSystem':
+        self.pipeline.fit(instances)
         return self
 
-    def apply(
-        self, instances: np.ndarray, labels: np.ndarray | None, meta: np.ndarray
-    ) -> tuple[np.ndarray, np.ndarray | None, np.ndarray]:
+    def apply(self, instances: FeatureData) -> LLRData:
         """
         Applies the specific source LR system on a set of instances, optionally with corresponding labels, and returns a
-        set of LLRs and their labels.
+        representation of the calculated LLR data through the `LLRData` tuple.
 
         The returned set of LLRs has the same order as the set of input instances, and the returned labels are unchanged
         from the input labels.
         """
         llrs = self.pipeline.transform(instances)
-        return llrs, labels, meta
+        return LLRData(**llrs.model_dump())
