@@ -15,18 +15,21 @@ See:
     Evaluating Forensic Science in the Courtroom, 2nd edition, 2016, pp. 96-97.
 """
 
+from typing import Any
+
 import matplotlib.pyplot as plt
 import numpy as np
 
 from lir.algorithms.isotonic_regression import IsotonicCalibrator
+from lir.data.models import LLRData
 from lir.util import (
-    odds_to_probability,
     logodds_to_odds,
+    odds_to_probability,
     probability_to_odds,
 )
 
 
-def plot_ece(llrs, labels, log_prior_odds_range=None, ax=plt):
+def plot_ece(llrdata: LLRData, log_prior_odds_range: tuple[float, float] = (-3, 3), ax: Any = plt) -> None:
     """
     Generates an ECE plot for a set of LRs and corresponding ground-truth
     labels.
@@ -42,26 +45,29 @@ def plot_ece(llrs, labels, log_prior_odds_range=None, ax=plt):
     :param log_prior_odds_range: the range of prior odds (tuple of two values,
         indicating both ends of the range on the x-axis)
     """
-    if log_prior_odds_range is None:
-        log_prior_odds_range = (-3, 3)
+    llrs = llrdata.llrs
+    labels = llrdata.labels
+    if labels is None:
+        raise ValueError('LLRData must contain labels to plot ECE.')
 
     log_prior_odds = np.arange(*log_prior_odds_range, 0.01)
     prior_odds = np.power(10, log_prior_odds)
 
     # plot reference
-    ax.plot(
-        log_prior_odds,
-        calculate_ece(np.ones(len(labels)), labels, odds_to_probability(prior_odds)),
-        linestyle=":",
-        label="reference",
-    )
+    if ax.get_legend() is None or 'reference' not in [text.get_text() for text in ax.get_legend().get_texts()]:
+        ax.plot(
+            log_prior_odds,
+            calculate_ece(np.ones_like(labels), labels, odds_to_probability(prior_odds)),
+            linestyle=':',
+            label='reference',
+        )
 
     # plot LRs
     ax.plot(
         log_prior_odds,
         calculate_ece(logodds_to_odds(llrs), labels, odds_to_probability(prior_odds)),
-        linestyle="-",
-        label="LRs",
+        linestyle='-',
+        label='LRs',
     )
 
     # plot PAV LRs
@@ -69,19 +75,19 @@ def plot_ece(llrs, labels, log_prior_odds_range=None, ax=plt):
     ax.plot(
         log_prior_odds,
         calculate_ece(logodds_to_odds(pav_llrs), labels, odds_to_probability(prior_odds)),
-        linestyle="--",
-        label="PAV LRs",
+        linestyle='--',
+        label='PAV LRs',
     )
 
-    ax.set_xlabel("prior log$_{10}$(odds)")
-    ax.set_ylabel("empirical cross-entropy")
+    ax.set_xlabel('prior log$_{10}$(odds)')
+    ax.set_ylabel('empirical cross-entropy')
     ax.set_ylim((0, None))
     ax.set_xlim(log_prior_odds_range)
     ax.legend()
-    ax.grid(True, linestyle=":")
+    ax.grid(True, linestyle=':')
 
 
-def calculate_ece(lrs, y, priors):
+def calculate_ece(lrs: np.ndarray, y: np.ndarray, priors: np.ndarray) -> np.ndarray:
     """
     Calculates the empirical cross-entropy (ECE) of a set of LRs and
     corresponding ground-truth labels.
@@ -95,14 +101,14 @@ def calculate_ece(lrs, y, priors):
         from class 1 (values in range [0..1])
     :returns: an array of entropy values of the same length as `priors`
     """
-    assert np.all(lrs >= 0), "invalid input for LR values"
-    assert np.all(np.unique(y) == np.array([0, 1])), "label set must be [0, 1]"
+    assert np.all(lrs >= 0), 'invalid input for LR values'
+    assert np.all(np.unique(y) == np.array([0, 1])), 'label set must be [0, 1]'
 
     prior_odds = np.repeat(probability_to_odds(priors), len(lrs)).reshape((len(priors), len(lrs)))
     posterior_odds = prior_odds * lrs
     posterior_p = odds_to_probability(posterior_odds)
 
-    with np.errstate(divide="ignore"):
+    with np.errstate(divide='ignore'):
         ece0 = -(1 - priors.reshape((len(priors), 1))) * np.log2(1 - posterior_p[:, y == 0])
         ece1 = -priors.reshape((len(priors), 1)) * np.log2(posterior_p[:, y == 1])
 
