@@ -1,21 +1,17 @@
-from collections.abc import Callable, Iterator, Sequence
 import csv
 import logging
-from abc import abstractmethod, ABC
+from abc import ABC, abstractmethod
+from collections.abc import Callable, Iterator
 from itertools import chain
 from pathlib import Path
-from typing import Any, Protocol
-from typing_extensions import Self
+from typing import Any, Protocol, Self
 
 import numpy as np
-from sklearn.base import TransformerMixin
+
+from lir.data.models import FeatureData
 
 
 LOG = logging.getLogger(__name__)
-
-TRANSFORMER_LABELS_KEY = "labels"
-TRANSFORMER_PHASE_KEY = "phase"
-TRANSFORMER_META_KEY = "meta"
 
 
 class DataWriter(Protocol):
@@ -28,12 +24,20 @@ class DataWriter(Protocol):
 class SKLearnPipelineModule(Protocol):
     """Representation of the interface required for estimators by the scikit-learn `Pipeline`."""
 
-    def fit(self, X: Sequence[Any], y: Sequence[Any] | None) -> Self: ...
-    def transform(self, X: Any) -> Any: ...
-    def predict_proba(self, X: Any) -> Any: ...
+    def fit(self, X: np.ndarray, y: np.ndarray | None) -> Self: ...
+    def transform(self, X: np.ndarray) -> Any: ...
+    def predict_proba(self, X: np.ndarray) -> Any: ...
 
 
-class Transformer(ABC, TransformerMixin):
+class SklearnTransformerType(Protocol):
+    """Representation of the interface required for transformers by the scikit-learn `Pipeline`."""
+
+    def fit(self, features: np.ndarray, labels: np.ndarray | None) -> Self: ...
+    def transform(self, features: np.ndarray) -> Any: ...
+    def fit_transform(self, features: np.ndarray, labels: np.ndarray | None) -> np.ndarray: ...
+
+
+class Transformer(ABC):
     """Transformer module which is compatible with the scikit-learn `Pipeline`.
 
     The transformer should provide a `transform()` method. Since transformers are not
@@ -41,28 +45,16 @@ class Transformer(ABC, TransformerMixin):
     side effects.
     """
 
-    def fit(self, X: Sequence[Any], y: Sequence[Any] | None = None) -> "Transformer":
+    def fit(self, instances: FeatureData) -> Self:
         return self
 
     @abstractmethod
-    def transform(self, X: Any) -> Any:
+    def transform(self, instances: FeatureData) -> Any:
         """Each transformer should implement a custom `transform()` method."""
         raise NotImplementedError
 
-
-class AdvancedTransformer(Transformer, ABC):
-    """Extended Transformer class which keeps internal state to be more flexible.
-
-    This type of transformer can be used if another "side effect" is desired while
-    performing a certain `Pipeline` step.
-    """
-
-    def __init__(self) -> None:
-        self._values: dict[str, Any] = {}
-
-    def set_value(self, key: str, value: Any | None) -> "AdvancedTransformer":
-        self._values[key] = value
-        return self
+    def fit_transform(self, instances: FeatureData) -> FeatureData:
+        return self.fit(instances).transform(instances)
 
 
 class Identity(Transformer):
@@ -76,12 +68,34 @@ class BinaryClassifierTransformer(Transformer):
     def __init__(self, estimator: SKLearnPipelineModule):
         self.estimator = estimator
 
-    def fit(self, X: Sequence[Any], y: Sequence[Any] | None = None) -> Transformer:
-        self.estimator.fit(X, y)
+    def fit(self, instances: FeatureData) -> Self:
+        self.estimator.fit(instances.features, instances.labels)
         return self
 
-    def transform(self, X: Sequence[Any]) -> Sequence[Any]:
-        return self.estimator.predict_proba(X)[:, 1]
+    def transform(self, instances: FeatureData) -> FeatureData:
+        # get probabilities from the estimator
+        probabilities = self.estimator.predict_proba(instances.features)[:, 1]
+        # return a copy of `instances` with the `features` attribute replaced by the newly obtained probabilities
+        return instances.replace(features=probabilities)
+
+
+class SklearnTransformer(Transformer):
+    """Implementation of a binary class classifier as scikit-learn `Pipeline` step."""
+
+    def __init__(self, transformer: SklearnTransformerType):
+        self.transformer = transformer
+
+    def fit(self, instances: FeatureData) -> Self:
+        self.transformer.fit(instances.features, instances.labels)
+        return self
+
+    def transform(self, instances: FeatureData) -> FeatureData:
+        return instances.replace_as(FeatureData, features=self.transformer.transform(instances.features))
+
+    def fit_transform(self, instances: FeatureData) -> FeatureData:
+        return instances.replace_as(
+            FeatureData, features=self.transformer.fit_transform(instances.features, instances.labels)
+        )
 
 
 class FunctionTransformer(Transformer):
@@ -90,42 +104,31 @@ class FunctionTransformer(Transformer):
     def __init__(self, func: Callable):
         self.func = func
 
-    def fit(self, X: Sequence[Any], y: Sequence[Any] | None = None) -> Transformer:
-        return self
-
-    def transform(self, X: Sequence[Any]) -> Sequence[Any]:
-        return self.func(X)
+    def transform(self, instances: FeatureData) -> FeatureData:
+        return instances.replace(features=self.func(instances.features))
 
 
-class Tee(AdvancedTransformer):
+class Tee(Transformer):
     """Implementation of a custom transformer allowing to perform two separate tasks on a given input."""
 
     def __init__(self, transformers: list[Transformer]):
         super().__init__()
         self.transformers = transformers
 
-    def set_value(self, key: str, value: Any) -> AdvancedTransformer:
-        super().set_value(key, value)
+    def fit(self, instances: FeatureData) -> Self:
         for transformer in self.transformers:
-            if isinstance(transformer, AdvancedTransformer):
-                transformer.set_value(key, value)
+            transformer.fit(instances)
 
         return self
 
-    def fit(self, X: Sequence[Any], y: Sequence[Any] | None = None) -> Transformer:
+    def transform(self, instances: FeatureData) -> FeatureData:
         for transformer in self.transformers:
-            transformer.fit(X, y)
+            transformer.transform(instances)
 
-        return self
-
-    def transform(self, X: Any) -> Any:
-        for transformer in self.transformers:
-            transformer.transform(X)
-
-        return X
+        return instances
 
 
-class TransformerWrapper(AdvancedTransformer):
+class TransformerWrapper(Transformer):
     """Base class for a transformer wrapper.
 
     This class is derived from `AdvancedTransformer` and has a default implementation of all functions
@@ -137,39 +140,35 @@ class TransformerWrapper(AdvancedTransformer):
         super().__init__()
         self.wrapped_transformer = wrapped_transformer
 
-    def set_value(self, key: str, value: Any | None) -> AdvancedTransformer:
-        super().set_value(key, value)
-        if isinstance(self.wrapped_transformer, AdvancedTransformer):
-            self.wrapped_transformer.set_value(key, value)
+    def fit(self, instances: FeatureData) -> Self:
+        self.wrapped_transformer.fit(instances)
         return self
 
-    def fit(self, X: Sequence[Any], y: Sequence[Any] | None = None) -> Transformer:
-        self.wrapped_transformer.fit(X, y)
-        return self
-
-    def transform(self, X: Any) -> Any:
-        return self.wrapped_transformer.transform(X)
+    def transform(self, instances: FeatureData) -> FeatureData:
+        return self.wrapped_transformer.transform(instances)
 
 
 class NumpyTransformer(TransformerWrapper):
-    """Implementation of a transformer wrapper to handle writing numpy data to CSV."""
+    """Implementation of a transformer wrapper."""
 
-    def __init__(self, transformer: Transformer, header: list[str] | None, path: Path):
+    def __init__(self, transformer: Transformer, header: list[str] | None):
         super().__init__(transformer)
-        self.csv_writer = CsvWriter(header=header, path=path)
+        self.header = header
 
-    def set_value(self, key: str, value: Any | None) -> AdvancedTransformer:
-        super().set_value(key, value)
-        self.csv_writer.set_value(key, value)
-        return self
+    def transform(self, instances: FeatureData) -> FeatureData:
+        instances = super().transform(instances)
+        if self.header:
+            instances = instances.replace(header=self.header)
+        return instances
 
-    def transform(self, X: Any) -> Any:
-        result = super().transform(X)
-        self.csv_writer.transform(result)
-        return result
+    def fit_transform(self, instances: FeatureData) -> FeatureData:
+        instances = super().fit_transform(instances)
+        if self.header:
+            instances = instances.replace(header=self.header)
+        return instances
 
 
-class CsvWriter(AdvancedTransformer):
+class CsvWriter(Transformer):
     """Implementation of a transformation step in a scikit-learn Pipeline that writes to CSV.
 
     This might be used to obtain temporary or intermediate results for logging or debugging
@@ -185,12 +184,11 @@ class CsvWriter(AdvancedTransformer):
         include_meta: bool = False,
         include_input: bool = True,
         include_batch: bool = False,
-        phase: str = "test",
     ):
         super().__init__()
         self.path = path
         if self.path is None:
-            raise ValueError("missing argument: path")
+            raise ValueError('missing argument: path')
 
         self.header_prefix = self.path.stem
 
@@ -200,83 +198,103 @@ class CsvWriter(AdvancedTransformer):
         self.include_meta = include_meta
         self.include_input = include_input
         self.include_batch = include_batch
-        self.write_on_phase = phase
         self.n_batches = 0
 
-    def _join_reader(self, filename: str) -> Iterator:
-        with open(self.path.parent / filename, "r") as f:
+    def _join_reader(self, filename: str) -> Iterator[list[str]]:
+        with open(self.path.parent / filename, 'r') as f:
             reader = csv.reader(f)
             yield from reader
 
-    def _write_rows(self, writer: DataWriter, rows: Any, write_header: bool) -> None:
-        all_headers = []
-        all_data = []
+    def _write_rows(self, writer: DataWriter, instances: FeatureData, write_header: bool) -> None:
+        all_headers: list[str] = []
+        all_data: list[np.ndarray] = []
 
-        if self.include_labels:
-            labels = self._values.get(TRANSFORMER_LABELS_KEY, None)
-            if labels is not None:
-                all_headers.append("label")
-                all_data.append(labels.reshape(-1, 1))
+        if self.include_labels and instances.labels is not None:
+            all_headers.append('label')
+            all_data.append(instances.labels.reshape(-1, 1))
 
-        if self.include_meta:
-            meta = self._values.get(TRANSFORMER_META_KEY, None)
-            if meta is not None:
-                if len(meta.shape) < 2:
-                    meta = meta.reshape(meta.shape[0], -1)
-                all_headers.extend([f"{self.header_prefix}.meta{n}" for n in range(meta.shape[1])])
-                all_data.append(meta)
+        if self.include_meta and hasattr(instances, 'meta'):
+            meta = instances.meta
+            if len(meta.shape) < 2:
+                meta = meta.reshape(meta.shape[0], -1)
+            all_headers.extend([f'{self.header_prefix}.meta{n}' for n in range(meta.shape[1])])
+            all_data.append(meta)
 
         if self.include_batch:
-            all_headers.append("batch")
-            all_data.append(np.ones((rows.shape[0], 1)) * self.n_batches)
+            all_headers.append('batch')
+            all_data.append(np.ones((len(instances), 1)) * self.n_batches)
 
         if self.include_input:
+            features = instances.features
+            if len(features.shape) == 1:
+                features = features.reshape(-1, 1)
+            elif len(features.shape) != 2:
+                raise ValueError(f'{__name__}: unsupported shape: {features.shape}')
+
             if self.header is None:
-                all_headers.extend([f"{self.header_prefix}{i}" for i in range(len(rows[0]))])
+                all_headers.extend([f'{self.header_prefix}{i}' for i in range(features.shape[1])])
             else:
                 all_headers.extend(self.header)
-            all_data.append(rows)
+            all_data.append(features)
 
         if self.input_files is not None:
             for filename in self.input_files:
                 rows = self._join_reader(filename)
                 all_headers.extend(next(rows))
-                all_data.append(rows)
+                all_data.append(np.array(list(rows)))
 
         if write_header:
             writer.writerow(all_headers)
-        for row in zip(*all_data):
+        for row in zip(*all_data, strict=True):
             writer.writerow(chain(*row))
 
-    def transform(self, features: np.ndarray) -> Any:
-        """Write numpy feature vector to CSV output file."""
-        if TRANSFORMER_PHASE_KEY not in self._values or self._values[TRANSFORMER_PHASE_KEY] != self.write_on_phase:
-            return features
+    def fit_transform(self, instances: FeatureData) -> FeatureData:
+        return instances
 
-        LOG.info(f"writing CSV file: {self.path}")
+    def transform(self, instances: FeatureData) -> FeatureData:
+        """Write numpy feature vector to CSV output file."""
+        LOG.info(f'writing CSV file: {self.path}')
         self.path.parent.mkdir(exist_ok=True, parents=True)
         if self.path.exists():
-            write_mode = "a"
+            write_mode = 'a'
             write_header = False
         else:
-            write_mode = "w"
+            write_mode = 'w'
             write_header = True
 
         with open(self.path, write_mode) as f:
             writer = csv.writer(f)
-
-            if isinstance(features, np.ndarray):
-                if len(features.shape) == 1:
-                    values = features.reshape(-1, 1)
-                elif len(features.shape) == 2:
-                    values = features
-                else:
-                    raise ValueError(f"{__name__}: unsupported shape: {features.shape}")
-                self._write_rows(writer, values, write_header)
-            else:
-                for row in features:
-                    writer.writerow(row)
+            self._write_rows(writer, instances, write_header)
 
         self.n_batches += 1
 
-        return features
+        return instances
+
+
+def as_transformer(transformer_like: Any) -> Transformer:
+    """
+    For any transformer-like object, wrap if necessary, and return a `Transformer`.
+
+    The transformer-like object may be one of the following:
+    - an instance of `Transformer`, which is returned as-is;
+    - a scikit-learn style transformer which implements `transform()` and optionally `fit()` and/or `fit_transform()`;
+    - a scikit-learn style estimator, which implements `fit()` and `predict_proba()`; or
+    - a callable which takes an `np.ndarray` argument and returns another `np.ndarray`.
+
+    :param transformer_like: the object to wrap or return
+    :return: a `Transformer` instance
+    """
+    if isinstance(transformer_like, Transformer):
+        # The component already supports all necessary methods, through the `Transformer` interface.
+        return transformer_like
+    if hasattr(transformer_like, 'transform'):
+        # The component implements a `transform()` method, which means we assume it is a sklearn style transformer.
+        return SklearnTransformer(transformer_like)
+    if hasattr(transformer_like, 'predict_proba'):
+        # The component has a `predict_proba` method, which means we assume it is a sklearn style estimator
+        return BinaryClassifierTransformer(transformer_like)
+    if callable(transformer_like):
+        # The component is a function
+        return FunctionTransformer(transformer_like)
+
+    raise ValueError(f'unknown module type of class {type(transformer_like)}')
